Use taskcluster-pulse to get credentials
This is backward-compatible: it's still possible to pass pulse
credentials directly, but the preferred approach is to pass taskcluster
credentials and let the library periodically re-claim the namespace.

NOTE: this modifies the exchange name used by the FakePublisher to
include `exchange/` and the namespace, which may cause tests to fail.